Move statusbar loghandler from own module to log
The module only contained the single class which fits very well into
utils.log. It was only imported there anyway.
